split retrofit service interface
